tests: move file-level describes into core main loop
We've been putting all content for each tests.js file into its own
describe block.  This makes it easy to use before/after hooks that
only run for tests defined in that file which is generally what we
expect to happen.  The downside to this is that the entire file has
to be indented one level to conform to the styleguide ... which we
have mostly been ignoring.  Let's move this logic into the main test
loop so we don't have to manually specify it and so we can start to
automate the formatting on JS files.

It means we can't define global before/after hooks in individual
_tests.js files, but we've never done that before, nor wanted to,
so we aren't losing anything.  If we truly want such hooks, they
can be placed in the xxx_test_main.js file which arguably makes it
a bit more discoverable & less surprising.

Performance wise, this seems to be about the same.

Change-Id: I0d6dc81b6ab2a13f72d3375ef6fb281711bfa0d5
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7273110
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>