jsdoc: duplicate @override content
Newer versions of closure-compiler require annotations be duplicated.
They changed behavior to help with internal TypeScript migration, but
included it with public releases.  Even if they fix it, newer versions
require newer JRE which we isn't available in Debian bookworm, and we
need to upgrade closure-compiler now for newer language support.

It at least verifies the duplicated types are the same, so if there's
a typo or one changes, it should warn.

On the upside, there seems to be cases where @override wasn't working
and types weren't being passed along & enforced.

Most of this was automated by loading the JSDoc from base classes and
replacing /** @override */ with that info.

Change-Id: I84eaaa80eb6d12007630286281c71d78ba792c3f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7228191
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>