wassh: handle_fd_pwrite: fix buf/offset naming swap
The order is supposed to be (fd, buffer, offset), and that's how the
wasi-js-bindings layer handles it, but wassh's handle_fd_pwrite had
them swapped.  The reason it worked anyways is that the VFS layer
had the order (buffer, offset), so wassh basically passes through
the 2nd & 3rd argument from wasi-js-bindings to the VFS, and the
name it uses here doesn't affect functional behavior.  We should
still fix it, especially as newer closure-compiler notices.

Change-Id: Icd56e3301bd2b08fa97edde7059f3de9186e09b3
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7220972
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>