nassh: command: use lib.Storage API consistently for sessionStorage
The CommandInstance declares its sessionStorage as lib.Storage, but it
assigns it from the default globalThis.sessionStorage which uses the
standard Storage web API which are not quite the same.  The former is
always async while the latter is always synchronous.  Fix the typing
so it's consistent, and fix the Relay types to match.

In practice, by using await on the APIs, it just works even if the
underlying API is actually synchronous.  So this change shouldn't be
functionally different.  Newer closure-compiler detects this mismatch.

Change-Id: I316f4cfc675e786b2d4bf35555ec77d6dc89c473
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7228190
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>