puppeteer: adjust how we disable downloads
Npm has deprecated project-specific configs in the .npmrc files.
As such, the only way to disable downloading is via the env.  Move
that logic to our node wrapper, and adjust the name as puppeteer
has deprecated the chromium-specific part.

The current versions of npm & puppeteer support both npmrc & env,
and the old & new names.  Switching now is the best time.

Change-Id: I1d2016b39be5a2d516d002cc0c95179348f912c6
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7234282
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>