terminal: tests: rename helper code to terminal_test_util.js
Since the canonical entry point is now "xxx_test_main.js", the old
"xxx_test.js" file has a bit of a naming crisis and isn't clear at
a glance what it's meant for, and it's easy to confuse with files
like "xxx_tests.js".  So let's adopt the "test_util.js" convention
that Chromium uses to indicate the file is full of code meant for
reuse among tests.

Change-Id: Ic18523deb22ba89f037790dac14d92200caf52d1
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7259740
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>