libdot: fetch: check cache based on uri not by the output
The DISTDIR stores artifacts using names based on the URI, so we need
to lookup the same way.  This has had been working previously because
we were saving files locally using the same name, but that changed in
the previous commit ("libdot: node: move download to fetch.json") as
it uses a simpler temporary name.

Change-Id: Idafed05b5059c6c264d119985104531442c51dd2
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7258514
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>