tests: undo one level of indentation after describe change
Now that we've unified the file-level suite grouping into the main
test loop and out of the actual tests files, we don't need to nest
the entire content of the files.  We weren't doing this already in
most places, but technically that was a style guide violation.

There are no changes here other than deleting 2 leading spaces.

Change-Id: I91ccb3bcc69b76774286b92366d92978fcf867db
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7282483
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>